+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - updates to the uk.json language file (#2436) (@xrust83)
 - updates to the th.json language file (#2437) (@tomzt)
 - updates to the `manual_update_command` language string
+- show restart modal after installing an already configured plugin
 
 ### Homebridge Dependencies